Added Notify/SharedNotify to unify message passing
There aren't many places where Cushy outputs a signal and doesn't expect
a response. Buttons and menus seem to be the only places where values
are sent and not received. Many of the value based widgets often need to
store and read the values, leading to channel integration not really
being useful.

So in the end, it seems like channel support in Cushy really is focused
around solving certain data flow problems easier by leveraging a runtime
provided by Cushy.