Fixed non_blocking channel callback edge case
When multiple values were in queue, the callback would get executed for
the first and then the future that called this method would only loop to
poll the generated futures before returning Poll::Pending. This would
leave the remaining values in queue.

Now the callback is repeatedly called until the queue is emptied.