Adding Runtime support to callbacks
One of the oversights of background execution of callbacks was ensuring
the cushy runtime was present. This forced me to ensure that only one
Cushy instance was installed at any given time, as the thread that
executes callbacks needs to enter the runtime. If this is set up before
the PendingApp is created, we need to ensure that the Cushy instance
PendingApp has and the one the callbacks have are the same one. By
ensuring we can only have one global instance, we meet this guarantee
and allow the casual user using the default runtime can create callbacks
before creating their PendingApp and everything works.