Added Default to various channels
A request from Discord was for Sender to implement Default. At first
glance, this seemed like a bad pattern to encourage, but as I thought
about it, the alternative for that use might be Option<Sender<T>>. It
would be significantly easier for a user to allow a disconnected sender
to determine if something can be sent or not. After that, it's a natural
conclusion to allow Default so that Sender<T> is as generally flexible
as Option<Sender<T>>.

This current implementation still allocates the structures for a channel
even though it can never end up being connected. We could optimize this
in the future at the cost of option checking on each send.