Fixing channel edge cases
The channel_disconnection test was flaky, and it was due to two separate
issues. First, when sending to a channel with the callback disconnected
while the future hadn't been disconnected fully yet, the send would
still be allowed rather than being reported as disconnected despite
never being able to be received. Second, using connected as a
distinction of whether the channel polling future should be enqueued was
not correct, as connected was being checked as to whether to poll for
notifications.