Background task queue overflow handling
The thread sender was limited to 256 pending change callbacks in queue.
A change callback went to enqueue when the channel was full, the
executor thread would deadlock.

Now, the executor thread uses its own queue for enqueuing change
callbacks, ensuring that it cannot be blocked due to too much inbound
traffic.