Fixed several issues found in examples
The containers example has slowed down at some point. Some very basic
benchmarking pointed to the font family list being applied being
expensive. I attempted to reduce the overhead of this by making single
element lists be allocation free. Unfortunately that wasn't enough, and
this slowdown is almost certainly unrelated to this branch, so while I'm
keeping the optimization, the example is still too slow.

In the process of trying to get baseline working for the Grid widget, at
one point I changed the "other" constraint to SizeToFit to prevent
labels from growing. Ultimately, the label code itself was updated to
account for this. By changing this back, many examples were working
correctly again.

Lastly, the selection rectangle for a single line of text wasn't
including the descender... which then became more complex since that
measurement isn't present if the text doesn't have any descenders.
There's an approximation in place now, but I suspect we should measure a
character with a descender to get the line metrics.